% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajv_methods.R
\name{ajv.addKeyword}
\alias{ajv.addKeyword}
\title{A wrapper for the Ajv.addKeyword method}
\usage{
ajv.addKeyword(this, name, definition)
}
\arguments{
\item{this}{An AJV instance, provided implicitly when called via \code{my_instance$addSchema(...)}}

\item{name}{The name of the keyword to be added.}

\item{definition}{A string encoding of a javascript object to be used as to define the keyword.}
}
\value{
\code{invisible(NULL)}
}
\description{
The add a schema to an Ajv instance
}
\examples{
\dontrun{
my_ajv_instance = Ajv()
my_ajv_instance$addSchema
}
}
\seealso{
Other AJV.Instance.Methods: \code{\link{ajv.addFormat}},
  \code{\link{ajv.addSchema}},
  \code{\link{ajv.errorsText}}, \code{\link{ajv.keyword}},
  \code{\link{ajv.removeSchema}},
  \code{\link{ajv.validateSchema}},
  \code{\link{ajv.validate}}
}

