% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_methods.R
\name{ajv.compile}
\alias{ajv.compile}
\title{The Ajv.compile method}
\usage{
ajv.compile(this, schema)
}
\arguments{
\item{this}{An AJV instance, provided implicitly when called via \code{my_instance$compile(...)}}

\item{schema}{The Schema with which to validate the \code{data}.
\code{schema} may be a  valid JSON string, an R object (i.e. `list(...)`), a
connection to a JSON file, or the name of a JSON or YAML file.  YAML files
are parsed via [js-yaml](https://www.npmjs.com/package/js-yaml)'s
`safeLoad()` method.}
}
\value{
an AJV validation function
}
\description{
Create an Ajv validator function from a schema
}

