% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajv_methods.R
\name{ajv.errorsText}
\alias{ajv.errorsText}
\title{A wrapper for the Ajv.errorsText method}
\usage{
ajv.errorsText(this)
}
\arguments{
\item{this}{An AJV instance, provided implicitly when called via \code{my_instance$errorsText(...)}}
}
\value{
JSON encoded object containing the error message (if any), with
class "AJV-errorsText" for pretty printing via \code{print.errorsText}
}
\description{
Extracts the errors object from
}
\examples{
\dontrun{
my_ajv_instance = Ajv()
my_ajv_instance$errorsText
}
}
\seealso{
Other AJV.Instance.Methods: \code{\link{ajv.addFormat}},
  \code{\link{ajv.addKeyword}},
  \code{\link{ajv.addSchema}}, \code{\link{ajv.keyword}},
  \code{\link{ajv.removeSchema}},
  \code{\link{ajv.validateSchema}},
  \code{\link{ajv.validate}}
}

