% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_methods.R
\name{ajv.getSchema}
\alias{ajv.getSchema}
\title{The Ajv.compile method}
\usage{
ajv.getSchema(this, key)
}
\arguments{
\item{this}{An AJV instance, provided implicitly when called via \code{my_instance$getSchema(...)}}

\item{key}{String; the name of the schema to fetch from the Ajv instancd.}
}
\value{
an AJV validation function
}
\description{
Create an Ajv validator function from a schema
}

