% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajv_methods.R
\name{ajv.keyword}
\alias{ajv.keyword}
\title{A wrapper for the Ajv.addFormat method}
\usage{
ajv.keyword(this, key, object)
}
\arguments{
\item{this}{An AJV instance, provided implicitly when called via \code{my_instance$keyword(...)}}

\item{key}{String; the name with keyword to add.}

\item{object}{the format to be added.  Must be enclosed in a call to
\code{\link[V8]{JS}}.}
}
\value{
\code{invisible(NULL)}
}
\description{
Add a string format to an Ajv instance.
}
\examples{
\dontrun{
my_ajv_instance = Ajv()
my_ajv_instance$keyword(key,object)
}
}
\seealso{
Other AJV.Instance.Methods: \code{\link{ajv.addFormat}},
  \code{\link{ajv.addKeyword}},
  \code{\link{ajv.addSchema}},
  \code{\link{ajv.errorsText}},
  \code{\link{ajv.removeSchema}},
  \code{\link{ajv.validateSchema}},
  \code{\link{ajv.validate}}
}

