% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ajv.R
\name{Ajv}
\alias{Ajv}
\title{Create an Ajv instnace.}
\usage{
Ajv(options = NULL)
}
\arguments{
\item{options}{Optional; see the \href{https://github.com/epoberezkin/ajv#options}{ajv github page} for details}
}
\description{
Create an Ajv instnace (the equivalent of calling \code{new Ajv()} in javascript)
}

